refactor page into own file
