+++ mediawiki/__init__.py
@@ -1,7 +1,8 @@
 '''
 mediawiki module initialization
 '''
-from .mediawiki import (MediaWiki, MediaWikiPage, URL, VERSION)
+from .mediawiki import (MediaWiki, URL, VERSION)
+from .mediawikipage import (MediaWikiPage)
 from .exceptions import (MediaWikiException, PageError, MediaWikiGeoCoordError,
                          RedirectError, DisambiguationError,
                          MediaWikiAPIURLError, HTTPTimeoutError,
